+++ 230417_TeamGitHub/Monster.cs
@@ -41,28 +41,5 @@ namespace _230417_TeamGitHub
     {
         // 좀비같이 생긴 좀비
     }
-    public class EliteMonster : Monster
-    {
-        // 해당 클래스를 상속한 몬스터는 더 강려크해진다
-        public virtual void Skill()
-        {
-            // 스킬을 쓴다. 오버라이딩 해서 무슨 스킬일지 정한다.
-        }
-    }
-    public class EliteOrc : EliteMonster
-    {
-        // 더 강한 오크. 근육이 늘었다. 그뉵그뉵.
-        public override void Skill()
-        {
-            // 분노분노분노하는 스킬
-        }
-    }
-    public class EliteZombie : EliteMonster
-    {
-        // 더 강한 좀비. 물리면 디버프에 걸린다.
-        public override void Skill()
-        {
-            // 이 좀비는 무해하지 않고 물지 않지 않습니다.
-        }
-    }
+
 }
